Add Account entity and update (fix) SQL migration for user_id type

+++ src/main/resources/db/migration/V1__create_user_and_account_tables.sql
@@ -8,7 +8,7 @@ CREATE TABLE users (
 
 CREATE TABLE accounts (
   id BIGSERIAL PRIMARY KEY, -- using BIGSERIAL to avoid issues with JPA long mapping.
-  user_id INTEGER NOT NULL REFERENCES users(id),
+  user_id BIGINT NOT NULL REFERENCES users(id),
   name VARCHAR(100) NOT NULL,
   type VARCHAR(20) NOT NULL,
   starting_balance NUMERIC(12,2) NOT NULL DEFAULT 0,
